+++ nuxeo-drive-client/tests/common.py
@@ -5,13 +5,9 @@ import hashlib
 import os
 import shutil
 import sys
-import tempfile
 import urllib2
-from unittest import TestCase
 
-import nxdrive
-from nxdrive.client import LocalClient, RemoteDocumentClient, \
-    RemoteFileSystemClient
+from nxdrive.client import RemoteDocumentClient
 from nxdrive.client.common import BaseClient
 from nxdrive.logging_config import configure, get_logger
 from nxdrive.utils import safe_long_path
@@ -78,6 +74,7 @@ def clean_dir(_dir):
     if test_data:
         shutil.move(to_remove, test_data)
         return
+
     try:
         for dirpath, folders, filenames in os.walk(to_remove):
             for folder in folders:
@@ -165,177 +162,3 @@ class RemoteDocumentClientForTests(RemoteDocumentClient):
 
     def result_set_query(self, query):
         return self.execute('Repository.ResultSetQuery', query=query)
-
-
-class IntegrationTestCase(TestCase):
-
-    def setUp(self):
-        # Save the current path for test files
-        self.location = dirname(__file__)
-
-        # Check the Nuxeo server test environment
-        self.nuxeo_url = os.environ.get('NXDRIVE_TEST_NUXEO_URL', 'http://localhost:8080/nuxeo')
-        self.admin_user = os.environ.get('NXDRIVE_TEST_USER', 'Administrator')
-        self.password = os.environ.get('NXDRIVE_TEST_PASSWORD', 'Administrator')
-        self.build_workspace = os.environ.get('WORKSPACE')
-
-        self.tmpdir = None
-        if self.build_workspace is not None:
-            self.tmpdir = os.path.join(self.build_workspace, "tmp")
-            if not os.path.isdir(self.tmpdir):
-                os.makedirs(self.tmpdir)
-
-        self.full_nuxeo_url = self.nuxeo_url
-        if '#' in self.nuxeo_url:
-            self.nuxeo_url = self.nuxeo_url.split('#')[0]
-        # Check the local filesystem test environment
-        self.local_test_folder_1 = tempfile.mkdtemp(u'drive-1', dir=self.tmpdir)
-        self.local_test_folder_2 = tempfile.mkdtemp(u'drive-2', dir=self.tmpdir)
-
-        self.local_nxdrive_folder_1 = os.path.join(
-            self.local_test_folder_1, u'Nuxeo Drive')
-        os.mkdir(self.local_nxdrive_folder_1)
-        self.local_nxdrive_folder_2 = os.path.join(
-            self.local_test_folder_2, u'Nuxeo Drive')
-        os.mkdir(self.local_nxdrive_folder_2)
-
-        self.nxdrive_conf_folder_1 = os.path.join(
-            self.local_test_folder_1, u'nuxeo-drive-conf')
-        os.mkdir(self.nxdrive_conf_folder_1)
-
-        self.nxdrive_conf_folder_2 = os.path.join(
-            self.local_test_folder_2, u'nuxeo-drive-conf')
-        os.mkdir(self.nxdrive_conf_folder_2)
-
-        self.version = nxdrive.__version__
-
-        # Long timeout for the root client that is responsible for the test
-        # environment set: this client is doing the first query on the Nuxeo
-        # server and might need to wait for a long time without failing for
-        # Nuxeo to finish initialize the repo on the first request after
-        # startup
-        root_remote_client = RemoteDocumentClientForTests(
-            self.nuxeo_url, self.admin_user,
-            u'nxdrive-test-administrator-device', self.version,
-            password=self.password, base_folder=u'/', timeout=60)
-
-        # Call the Nuxeo operation to setup the integration test environment
-        credentials = root_remote_client.execute(
-            "NuxeoDrive.SetupIntegrationTests",
-            userNames="user_1, user_2", permission='ReadWrite')
-
-        credentials = [c.strip().split(u":") for c in credentials.split(u",")]
-        self.user_1, self.password_1 = credentials[0]
-        self.user_2, self.password_2 = credentials[1]
-
-        ws_info = root_remote_client.fetch(TEST_WORKSPACE_PATH)
-        self.workspace = ws_info[u'uid']
-        self.workspace_title = ws_info[u'title']
-
-        # Document client to be used to create remote test documents
-        # and folders
-        self.upload_tmp_dir = tempfile.mkdtemp(u'-nxdrive-uploads', dir=self.tmpdir)
-        remote_document_client_1 = RemoteDocumentClientForTests(
-            self.nuxeo_url, self.user_1, u'nxdrive-test-device-1',
-            self.version,
-            password=self.password_1, base_folder=self.workspace,
-            upload_tmp_dir=self.upload_tmp_dir)
-
-        remote_document_client_2 = RemoteDocumentClientForTests(
-            self.nuxeo_url, self.user_2, u'nxdrive-test-device-2',
-            self.version,
-            password=self.password_2, base_folder=self.workspace,
-            upload_tmp_dir=self.upload_tmp_dir)
-
-        # File system client to be used to create remote test documents
-        # and folders
-        remote_file_system_client_1 = RemoteFileSystemClient(
-            self.nuxeo_url, self.user_1, u'nxdrive-test-device-1',
-            self.version,
-            password=self.password_1, upload_tmp_dir=self.upload_tmp_dir)
-
-        remote_file_system_client_2 = RemoteFileSystemClient(
-            self.nuxeo_url, self.user_2, u'nxdrive-test-device-2',
-            self.version,
-            password=self.password_2, upload_tmp_dir=self.upload_tmp_dir)
-
-        self.root_remote_client = root_remote_client
-        self.remote_document_client_1 = remote_document_client_1
-        self.remote_document_client_2 = remote_document_client_2
-        self.remote_file_system_client_1 = remote_file_system_client_1
-        self.remote_file_system_client_2 = remote_file_system_client_2
-
-        self.local_client_1 = LocalClient(os.path.join(self.local_nxdrive_folder_1, self.workspace_title))
-        self.local_client_2 = LocalClient(os.path.join(self.local_nxdrive_folder_2, self.workspace_title))
-
-        self.ndrive_exec = os.path.join(self.location, '..', 'scripts', 'ndrive')
-        cmdline_options = ' --log-level-file=TRACE'
-        cmdline_options += ' --nxdrive-home="%s"'
-        if os.environ.get('PYDEV_DEBUG') == 'True':
-            cmdline_options += ' --debug-pydev'
-        self.ndrive_1_options = cmdline_options % self.nxdrive_conf_folder_1
-        self.ndrive_2_options = cmdline_options % self.nxdrive_conf_folder_2
-
-    def tearDown(self):
-        # Don't need to revoke tokens for the file system remote clients
-        # since they use the same users as the remote document clients
-        try:
-            self.root_remote_client.execute("NuxeoDrive.TearDownIntegrationTests")
-        except AttributeError:
-            # If a test did not have enough time, failed early, `root_remote_client` could be inexistant. Just ignore.
-            pass
-
-        clean_dir(self.upload_tmp_dir)
-        clean_dir(self.local_test_folder_1)
-        clean_dir(self.local_test_folder_2)
-
-    def wait(self, retry=3):
-        try:
-            self.root_remote_client.wait()
-        except Exception as e:
-            log.debug("Exception while waiting for server : %r", e)
-            # Not the nicest
-            if retry > 0:
-                log.debug("Retry to wait")
-                self.wait(retry - 1)
-
-    def setUpDrive_1(self, bind_root=True, root=None, first_sync=False):
-        # Bind the server and root workspace
-        self.bind_server(self.ndrive_1_options, self.user_1, self.nuxeo_url, self.local_nxdrive_folder_1,
-                         self.password_1)
-        if bind_root:
-            root_to_bind = root if root is not None else self.workspace
-            self.bind_root(self.ndrive_1_options, root_to_bind, self.local_nxdrive_folder_1)
-        if first_sync:
-            self.ndrive(self.ndrive_1_options)
-
-    def bind_root(self, ndrive_options, workspace, local_folder):
-        cmdline = '%s bind-root "%s" --local-folder="%s" %s' % (self.ndrive_exec, workspace, local_folder,
-                                                                ndrive_options)
-        execute(cmdline)
-
-    def unbind_root(self, ndrive_options, workspace, local_folder):
-        cmdline = '%s unbind-root "%s" --local-folder="%s" %s' % (self.ndrive_exec, workspace, local_folder,
-                                                                  ndrive_options)
-        execute(cmdline)
-
-    def bind_server(self, ndrive_options, user, server_url, local_folder, password):
-        cmdline = '%s bind-server %s %s --local-folder="%s" --password=%s %s' % (
-            self.ndrive_exec, user, server_url, local_folder, password, ndrive_options)
-        execute(cmdline)
-
-    def unbind_server(self, ndrive_options, local_folder):
-        cmdline = '%s unbind-server --local-folder="%s" %s' % (self.ndrive_exec, local_folder, ndrive_options)
-        execute(cmdline)
-
-    def ndrive(self, ndrive_options=None, quit_if_done=True, quit_timeout=TEST_DEFAULT_QUIT_TIMEOUT,
-               delay=TEST_DEFAULT_DELAY):
-        if not ndrive_options:
-            ndrive_options = self.ndrive_1_options
-        cmdline = self.ndrive_exec + ' console'
-        if quit_if_done:
-            cmdline += ' --quit-if-done'
-        cmdline += ' --quit-timeout=%d' % quit_timeout
-        cmdline += ' --delay=%d' % delay
-        cmdline += ' ' + ndrive_options
-        execute(cmdline)